requirements.txt
pandas
sqlalchemy
pymysql
python-dotenv
tabulate
jinja2
